// web/src/api/generated/models/ServerAdminSpaceDetail.ts
export interface ServerAdminSpaceDetail {
    /**
     * 
     * @type {number}
     * @memberof ServerAdminSpaceDetail
     */
    activeMembershipCount: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof ServerAdminSpaceDetail
     */
    anomalyCodes: Array<string>;
    /**
     * 
     * @type {Date}
     * @memberof ServerAdminSpaceDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ServerAdminSpaceDetail
     */
    firstMembershipAt: Date | null;
    /**
     * 
     * @type {number}
     * @memberof ServerAdminSpaceDetail
     */
    historicalMembershipCount: number;
    /**
     * 
     * @type {string}
     * @memberof ServerAdminSpaceDetail
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof ServerAdminSpaceDetail
     */
    lastMembershipChangeAt: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof ServerAdminSpaceDetail
     */
    latestMembershipEndedAt: Date | null;
    /**
     * 
     * @type {number}
     * @memberof ServerAdminSpaceDetail
     */
    leftMembershipCount: number;
    /**
     * 
     * @type {ServerAdminSpaceDetailLifecycleStatusEnum}
     * @memberof ServerAdminSpaceDetail
     */
    lifecycleStatus: ServerAdminSpaceDetailLifecycleStatusEnum;
    /**
     * 
     * @type {number}
     * @memberof ServerAdminSpaceDetail
     */
    membershipCount: number;
    /**
     * 
     * @type {number}
     * @memberof ServerAdminSpaceDetail
     */
    removedMembershipCount: number;
}


/**
 * @export
 */
export const ServerAdminSpaceDetailLifecycleStatusEnum = {
    active: 'active',
    inactive: 'inactive',
    empty: 'empty'
} as const;
export type ServerAdminSpaceDetailLifecycleStatusEnum = typeof ServerAdminSpaceDetailLifecycleStatusEnum[keyof typeof ServerAdminSpaceDetailLifecycleStatusEnum];